+++ cdk/kraken/.projenrc.js
@@ -4,7 +4,7 @@ const common = require('../projen-common');
 const project = new TypeScriptProject({
   name: '@pennlabs/kraken',
   description: 'cdkactions construct abstractions built for Penn Labs',
-  deps: ['cdkactions', 'constructs', 'dedent-js'],
+  deps: ['cdkactions', 'constructs', 'ts-dedent'],
   devDeps: ['codecov'],
   keywords: ['cdk', 'github', 'actions', 'constructs', 'cdkactions'],
   homepage: 'https://kraken.pennlabs.org',