+++ cdk/kraken/src/pypi.ts
@@ -1,5 +1,6 @@
 import { CheckoutJob, Workflow, Stack, WorkflowProps } from 'cdkactions';
 import { Construct } from 'constructs';
+import dedent from 'ts-dedent';
 
 /**
  * Optional props to configure the PyPI publish stack.
@@ -10,6 +11,12 @@ export interface PyPIPublishStackProps {
    * @default "3.8"
    */
   pythonVersion?: string;
+
+  /**
+   * List of python versions to run tox with.
+   * @default "3.7", "3.8", "3.9", "3.10"
+   */
+  pythonMatrixVersions?: string[];
 }
 
 /**
@@ -27,6 +34,7 @@ export class PyPIPublishStack extends Stack {
     // Build config
     const fullConfig: Required<PyPIPublishStackProps> = {
       pythonVersion: '3.8',
+      pythonMatrixVersions: ['3.7', '3.8', '3.9', '3.10'],
       ...config,
     };
 
@@ -41,15 +49,24 @@ export class PyPIPublishStack extends Stack {
       },
       ...overrides,
     });
-    new CheckoutJob(workflow, 'test', {
+    const testJob = new CheckoutJob(workflow, 'test', {
       runsOn: 'ubuntu-latest',
-      container: {
-        image: 'themattrix/tox',
+      strategy: {
+        matrix: {
+          'python-version': fullConfig.pythonMatrixVersions,
+        },
       },
       steps: [
         {
+          name: 'Set up Python ${{ matrix.python-version }}',
+          uses: 'actions/setup-python@v2',
+          with: {
+            'python-version': '${{ matrix.python-version }}',
+          },
+        },
+        {
           name: 'Install dependencies',
-          run: 'pip install codecov',
+          run: 'pip install poetry tox tox-gh-actions codecov',
         },
         {
           name: 'Test',
@@ -67,22 +84,29 @@ export class PyPIPublishStack extends Stack {
       container: {
         image: `python:${fullConfig.pythonVersion}`,
       },
+      needs: testJob.id,
       if: "startsWith(github.ref, 'refs/tags')",
       steps: [
         {
+          name: 'Install dependencies',
+          run: 'pip install poetry',
+        },
+        {
           name: 'Verify tag',
-          run: 'python3 setup.py verify',
+          shell: 'bash',
+          run: dedent`GIT_TAG=\${GITHUB_REF/refs\\/tags\\//}
+          LIBRARY_VERSION=$(poetry version -s)
+          if [[ "$GIT_TAG" != "$LIBRARY_VERSION" ]]; then echo "Tag ($GIT_TAG) does not match poetry version ($LIBRARY_VERSION)"; exit 1; fi`,
         },
         {
           name: 'Build',
-          run: 'python3 setup.py sdist bdist_wheel',
+          run: 'poetry build',
         },
         {
           name: 'Publish',
-          uses: 'pypa/gh-action-pypi-publish@v1',
-          with: {
-            user: '__token__',
-            password: '${{ secrets.PYPI_PASSWORD }}',
+          run: 'poetry publish',
+          env: {
+            POETRY_PYPI_TOKEN_PYPI: '${{ secrets.PYPI_PASSWORD }}',
           },
         },
       ],