+++ cdk/kraken/CHANGELOG.md
@@ -1,8 +1,38 @@
 # Changelog
 
-## 0.6.0 (UNRELEASED)
+## X.Y.Z (UNRELEASED)
 
 * Migrate to kittyhawk for deployments
+
+## 0.7.1 (2021-12-16)
+
+* Modify PyPI to correctly publish (again)
+
+## 0.7.0 (2021-12-04)
+
+* Modify PyPI python versions to be strings
+* Modify PyPI to correctly publish
+* Migrate to ts-dedent
+
+## 0.6.4 (2021-10-04)
+
+* Modify PyPI publish to use poetry
+
+## 0.6.3 (2021-08-26)
+
+* Pin OS to buster within django check
+
+## 0.6.2 (2021-04-04)
+
+* Add poetry install to PyPI stack
+
+## 0.6.1 (2021-04-04)
+
+* Fix PyPI job dependencies
+
+## 0.6.0 (2021-04-01)
+
+* Modify PyPI stack to use a matrix
 
 ## 0.5.1 (2021-03-17)